Did file writing stuff (READ DESCRIPTION)
I made the writeAccToFile(); methods that take either a savings account or a checking account and will write them into the file. 
To whoever is making the code to collect data from the user: take that user data and make a bank account object out of it. 

When you make the bank account, RUN writeAccToFile(SavingsAccount s) or writeAccToFile(CheckingAccount c) to add that account. 

- Also, when the program creates the file it now displays the date of creation at the top of the file.